Added protection for routes, renamed some files for better organization
